// src/app/models/homeConfig.model.ts
export interface HomeConfig{
    id?: number
    slider1?: string;
    title1?: string;
    text1?: string;
    slider2?: string;
    title2?: string;
    text2?: string;
    slider3?: string;
    title3?: string;
    text3?: string;
    cervezasCatalog?: string;
    vinosCatalog?: string;
    regalosCatalog?: string;
    whiskysCatalog?: string;
    combosCatalog?: string;
    aperitivosCatalog?: string;
    destiladosCatalog?: string;
    sinAlcoholCatalog?: string;
    elegidoDelMes?: string;
    masVendido1?: string;
    masVendido2?: string;
    masVendido3?: string;
    masVendido4?: string;
    spronsors?: string;

//     constructor(id: number,
//         slider1: string,
//         title1:string,
//         text1: string,
//         slider2: string,
//         title2: string,
//         text2: string,
//         slider3: string,
//         title3: string,
//         text3:string,
//         cervezasCatalog:string,
//         aperitivosCatalog:string,
//         vinosCatalog:string,
//         regalosCatalog:string,
//         whiskysCataglog:string,
//         combosCatalog:string,
//         destiladosCatalog:string,
//         sinAlcoholCatalog:string,
//         elegidoDelMes:string,
//         masVendido1:string,
//         masVendido2:string,
//         masVentido3:string,
//         masVendido4:string,
//         sponsors:string,
    
//         ) {
            
// this.id = id;
// this.slider1 = slider1;
// this.title1 = title1;
// this.text1 = text1;
// this.slider2 = slider2;
// this.title2 = title2;
// this.text2 = text2;
// this.slider3 = slider3;
// this.title3 = title3;
// this.text3 = text3;
// this.cervezasCatalog = cervezasCatalog;
// this.vinosCatalog = vinosCatalog;
// this.aperitivosCatalog = aperitivosCatalog;
// this.regalosCatalog = regalosCatalog;
// this.whiskysCatalog = whiskysCataglog;
// this.combosCatalog = combosCatalog;
// this.destiladosCatalog = destiladosCatalog;
// this.sinAlcoholCatalog = sinAlcoholCatalog;
// this.elegidoDelMes = elegidoDelMes;
// this.masVendido1 = masVendido1;
// this.masVendido2 = masVendido2;
// this.masVendido3 = masVentido3;
// this.masVendido4 = masVendido4;
// this.spronsors = sponsors;
//}
}